+++ src/Redux/ConfigureStore.js
@@ -2,9 +2,9 @@ import { configureStore } from "@reduxjs/toolkit";
 import { combineReducers, applyMiddleware } from "redux";
 import thunk from "redux-thunk";
 import logger from "redux-logger";
-import BrandReducer from "./makeupReducer";
+import BrandReducer from "./BrandReducer";
 
-const RootReducer = combineReducers({ Brand: BrandReducer });
+const RootReducer = combineReducers({ BrandReducer });
 
 const Store = configureStore(
   { reducer: RootReducer },
