+++ src/components/pages/Homepage.js
@@ -1,4 +1,15 @@
-import React from 'react';
+import React, { useEffect } from "react";
+import { useSelector, useDispatch } from "react-redux";
+import { BrandStore } from "../../Redux/BrandReducer";
 
-const HomePage = () => <div>Homepage</div>;
+const HomePage = () => {
+  const dispatch = useDispatch();
+  const BrandNames = useSelector((state) => state);
+  console.log(BrandNames);
+
+  useEffect(() => {
+    dispatch(BrandStore());
+  }, []);
+  return <div>HomePage</div>;
+};
 export default HomePage;
